Refactor catalog schemas to streamline component structure
- Updated import paths for schema files to improve organization and clarity.
- Introduced new `DescriptionList`, `Header`, `List`, `CardStacked`, and `ListStandard` schemas to enhance component functionality and maintainability.
- Consolidated existing schemas and removed deprecated components to simplify the catalog structure.
- Enhanced properties across various schemas for improved layout control and visual customization.
